\name{AUC.emp}
\alias{AUC.emp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function for AUC when input is X and Y
}
\description{
NA
}
\usage{
AUC.emp(X, Y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
    m X p data matrix for m non-diseased subjects with p markers
}
  \item{Y}{
    n X p data matrix for n diseased subjects with p markers
}
}

\author{
Maintainer: Xin Huang <xhuang.fhcrc@gmail.com>
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
