\name{beta2theta}
\alias{beta2theta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to translate beta into theta, the n-sphere constrain 
}
\description{
Function to translate beta into theta, the n-sphere constrain 
}
\usage{
beta2theta(beta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
   coefficients for linear combination of multiple diagnostic tests
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
