\name{gradAUC.Lang}
\alias{gradAUC.Lang}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function for gradient of AUC after applying Lagrange Multiplyer
}
\description{
Function for gradient of AUC after applying Lagrange Multiplyer
}
\usage{
gradAUC.Lang(par, Z, lambda)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par}{
paramter
}
  \item{Z}{
Z
}
  \item{lambda}{
the smooth parameter for the Sigmoid function used for the AUC 
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
