\name{nlsolve}
\alias{nlsolve}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
solve nonlinear function
}
\description{
solve nonlinear function
}
\usage{
nlsolve(par, fn, method = "BFGS", nstarts = 1, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par}{
parameter
}
  \item{fn}{
function
}
  \item{method}{
method
}
  \item{nstarts}{
tries
}
  \item{\dots}{
other paramters pass to fn
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
