\name{theta2beta}
\alias{theta2beta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to translate theta to beta, the n-sphere constrain
}
\description{
Function to translate theta to beta, the n-sphere constrain
}
\usage{
theta2beta(theta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theta}{
the parameter on n unit-sphere constrain
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
